Refactor of clipboard to separate out file saving/loading
There's now a clipboard manager to handle file system loading and saving
so that the clipboard works purely in blocks (YAML).
