+++ src/main/java/world/bentobox/bentobox/api/commands/admin/schem/AdminSchemCommand.java
@@ -13,8 +13,8 @@ import org.bukkit.Particle;
 import world.bentobox.bentobox.api.commands.CompositeCommand;
 import world.bentobox.bentobox.api.commands.ConfirmableCommand;
 import world.bentobox.bentobox.api.user.User;
+import world.bentobox.bentobox.blueprints.Clipboard;
 import world.bentobox.bentobox.managers.SchemsManager;
-import world.bentobox.bentobox.schems.Clipboard;
 
 public class AdminSchemCommand extends ConfirmableCommand {
     // Clipboards
