+++ src/main/java/world/bentobox/bentobox/api/commands/admin/schem/AdminSchemOriginCommand.java
@@ -8,7 +8,7 @@ import org.bukkit.block.Block;
 
 import world.bentobox.bentobox.api.commands.CompositeCommand;
 import world.bentobox.bentobox.api.user.User;
-import world.bentobox.bentobox.schems.Clipboard;
+import world.bentobox.bentobox.blueprints.Clipboard;
 
 public class AdminSchemOriginCommand extends CompositeCommand {
 
@@ -26,7 +26,7 @@ public class AdminSchemOriginCommand extends CompositeCommand {
     public boolean execute(User user, String label, List<String> args) {
         AdminSchemCommand parent = (AdminSchemCommand) getParent();
 
-        Clipboard clipboard = parent.getClipboards().getOrDefault(user.getUniqueId(), new Clipboard(getPlugin(), parent.getSchemsFolder()));
+        Clipboard clipboard = parent.getClipboards().computeIfAbsent(user.getUniqueId(), v -> new Clipboard());
         if (clipboard.getPos1() == null || clipboard.getPos2() == null) {
             user.sendMessage("commands.admin.schem.need-pos1-pos2");
             return false;
