+++ src/main/java/world/bentobox/bentobox/api/commands/admin/schem/AdminSchemPasteCommand.java
@@ -4,7 +4,8 @@ import java.util.List;
 
 import world.bentobox.bentobox.api.commands.CompositeCommand;
 import world.bentobox.bentobox.api.user.User;
-import world.bentobox.bentobox.schems.Clipboard;
+import world.bentobox.bentobox.blueprints.Clipboard;
+import world.bentobox.bentobox.blueprints.Paster;
 
 public class AdminSchemPasteCommand extends CompositeCommand {
 
@@ -22,10 +23,10 @@ public class AdminSchemPasteCommand extends CompositeCommand {
     public boolean execute(User user, String label, List<String> args) {
         AdminSchemCommand parent = (AdminSchemCommand) getParent();
 
-        Clipboard clipboard = parent.getClipboards().getOrDefault(user.getUniqueId(), new Clipboard(getPlugin(), parent.getSchemsFolder()));
+        Clipboard clipboard = parent.getClipboards().computeIfAbsent(user.getUniqueId(), v -> new Clipboard());
         if (clipboard.isFull()) {
-            clipboard.pasteClipboard(user.getLocation());
-            user.sendMessage("general.success");
+            new Paster(getPlugin(), clipboard, user.getLocation(), () -> user.sendMessage("general.success"));
+            user.sendMessage("commands.admin.schem.paste.pasting");
             return true;
         }
