+++ src/main/java/world/bentobox/bentobox/managers/ClipboardManager.java
@@ -1,0 +1,208 @@
+package world.bentobox.bentobox.managers;
+
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
+import java.util.zip.ZipOutputStream;
+
+import org.bukkit.Location;
+import org.bukkit.World;
+import org.bukkit.configuration.InvalidConfigurationException;
+import org.bukkit.configuration.file.YamlConfiguration;
+
+import world.bentobox.bentobox.BentoBox;
+import world.bentobox.bentobox.api.user.User;
+import world.bentobox.bentobox.blueprints.Clipboard;
+import world.bentobox.bentobox.blueprints.Paster;
+import world.bentobox.bentobox.database.objects.Island;
+
+/**
+ * @author tastybento
+ */
+public class ClipboardManager {
+
+    private static final String LOAD_ERROR = "Could not load schems file - does not exist : ";
+
+    private BentoBox plugin;
+
+    private File schemFolder;
+
+    private Clipboard clipboard;
+
+    public ClipboardManager(BentoBox plugin, File schemFolder) {
+        this(plugin, schemFolder, null);
+    }
+
+    public ClipboardManager(BentoBox plugin, File schemFolder, Clipboard clipboard) {
+        super();
+        this.plugin = plugin;
+        if (!schemFolder.exists()) {
+            schemFolder.mkdirs();
+        }
+        this.schemFolder = schemFolder;
+        this.clipboard = clipboard;
+    }
+
+    private void unzip(final String zipFilePath) throws IOException {
+        Path path = Paths.get(zipFilePath);
+        if (!(path.toFile().exists())) {
+            throw new IOException("No file exists!");
+        }
+        try (ZipInputStream zipInputStream = new ZipInputStream(new FileInputStream(zipFilePath))) {
+            ZipEntry entry = zipInputStream.getNextEntry();
+            while (entry != null) {
+                Path filePath = Paths.get(path.getParent().toString(), entry.getName());
+                if (!entry.isDirectory()) {
+                    unzipFiles(zipInputStream, filePath);
+                } else {
+                    Files.createDirectories(filePath);
+                }
+
+                zipInputStream.closeEntry();
+                entry = zipInputStream.getNextEntry();
+            }
+        }
+    }
+
+    private static void unzipFiles(final ZipInputStream zipInputStream, final Path unzipFilePath) throws IOException {
+
+        try (BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(unzipFilePath.toAbsolutePath().toString()))) {
+            byte[] bytesIn = new byte[1024];
+            int read;
+            while ((read = zipInputStream.read(bytesIn)) != -1) {
+                bos.write(bytesIn, 0, read);
+            }
+        }
+
+    }
+
+    private void zip(File targetFile) throws IOException {
+        try (ZipOutputStream zipOutputStream = new ZipOutputStream(new FileOutputStream(targetFile.getAbsolutePath() + ".schem"))) {
+            zipOutputStream.putNextEntry(new ZipEntry(targetFile.getName()));
+            try (FileInputStream inputStream = new FileInputStream(targetFile)) {
+                final byte[] buffer = new byte[1024];
+                int length;
+                while((length = inputStream.read(buffer)) >= 0) {
+                    zipOutputStream.write(buffer, 0, length);
+                }
+                try {
+                    Files.delete(targetFile.toPath());
+                } catch (Exception e) {
+                    plugin.logError(e.getMessage());
+                }
+            }
+        }
+    }
+
+    /**
+     * Load a file to clipboard
+     * @param fileName - filename in schems folder
+     * @throws IOException - if there's a load error with unziping or name
+     * @throws InvalidConfigurationException - the YAML of the schem is at fault
+     */
+    public void load(String fileName) throws IOException, InvalidConfigurationException {
+        File zipFile = new File(schemFolder, fileName + ".schem");
+        if (!zipFile.exists()) {
+            plugin.logError(LOAD_ERROR + zipFile.getName());
+            throw new IOException(LOAD_ERROR + zipFile.getName());
+        }
+        unzip(zipFile.getAbsolutePath());
+        File file = new File(schemFolder, fileName);
+        if (!file.exists()) {
+            plugin.logError(LOAD_ERROR + file.getName());
+            throw new IOException(LOAD_ERROR + file.getName());
+        }
+
+        YamlConfiguration blockConfig = new YamlConfiguration();
+        blockConfig.load(file);
+        clipboard = new Clipboard(blockConfig);
+        Files.delete(file.toPath());
+    }
+
+    /*
+      Load a file to clipboard
+     */
+    /**
+     * @param user - user trying to load
+     * @param fileName - filename
+     * @return - <tt>true</tt> if load is successful, <tt>false</tt> if not
+     */
+    public boolean load(User user, String fileName) {
+        try {
+            load(fileName);
+        } catch (IOException e1) {
+            user.sendMessage("commands.admin.schem.could-not-load");
+            plugin.logError("Could not load schems file: " + fileName + " " + e1.getMessage());
+            return false;
+        } catch (InvalidConfigurationException e1) {
+            user.sendMessage("commands.admin.schem.could-not-load");
+            plugin.logError("Could not load schems file - YAML error : " + fileName + " " + e1.getMessage());
+            return false;
+        }
+        user.sendMessage("general.success");
+        return true;
+    }
+
+    /**
+     * Save the clipboard to a file
+     * @param user - user who is copying
+     * @param newFile - filename
+     * @return - true if successful, false if error
+     */
+    public boolean save(User user, String newFile) {
+        File file = new File(schemFolder, newFile);
+        try {
+            clipboard.getBlockConfig().save(file);
+        } catch (IOException e) {
+            user.sendMessage("commands.admin.schem.could-not-save", "[message]", "Could not save temp schems file.");
+            plugin.logError("Could not save temporary schems file: " + file.getName());
+            return false;
+        }
+        try {
+            zip(file);
+        } catch (IOException e) {
+            user.sendMessage("commands.admin.schem.could-not-save", "[message]", "Could not zip temp schems file.");
+            plugin.logError("Could not zip temporary schems file: " + file.getName());
+            return false;
+        }
+        user.sendMessage("general.success");
+        return true;
+    }
+
+    /**
+     * Paste clipboard to this location
+     * @param location - location
+     */
+    public void pasteClipboard(Location location) {
+        if (clipboard != null) {
+            new Paster(plugin, clipboard, location);
+        } else {
+            plugin.logError("Clipboard has no block data in it to paste!");
+        }
+    }
+    /**
+     * Pastes the clipboard to island location.
+     * If pos1 and pos2 are not set already, they are automatically set to the pasted coordinates
+     * @param world - world in which to paste
+     * @param island - location to paste
+     * @param task - task to run after pasting
+     */
+    public void pasteIsland(World world, Island island, Runnable task) {
+        new Paster(plugin, clipboard, world, island, task);
+    }
+
+    /**
+     * @return the clipboard
+     */
+    public Clipboard getClipBoard() {
+        return clipboard;
+    }
+
+}
